handlerのファイル配置変更と警告解除
