+++ samples/web-csr/dressca-frontend/admin/mock/handlers/index.ts
@@ -1,0 +1,13 @@
+import { catalogItemsHandlers } from './catalog-items-handler';
+import { catalogBrandsHandlers } from './catalog-brands-handler';
+import { catalogCategoriesHandlers } from './catalog-categories-handler';
+import { assetsHandlers } from './assets-handler';
+import { usersHandlers } from './users-handler';
+
+export const handlers = [
+  ...catalogItemsHandlers,
+  ...catalogBrandsHandlers,
+  ...catalogCategoriesHandlers,
+  ...assetsHandlers,
+  ...usersHandlers,
+];
