+++ samples/web-csr/dressca-frontend/consumer/mock/handlers/baskets-handler.ts
@@ -10,13 +10,16 @@ function calcBasketAccount() {
   if (!basket || !basket.account) {
     return;
   }
-  const totalItemsPrice = basket.basketItems
-    ?.map((item) => {
-      // eslint-disable-next-line no-param-reassign
-      item.subTotal = item.unitPrice * item.quantity;
-      return item.subTotal;
-    })
-    .reduce((acc, subTotal) => acc + subTotal, 0);
+  // undefined になる場合は 0 を代入
+  const totalItemsPrice = basket.basketItems?.length
+    ? basket.basketItems
+        .map((item) => {
+          // eslint-disable-next-line no-param-reassign
+          item.subTotal = item.unitPrice * item.quantity;
+          return item.subTotal;
+        })
+        .reduce((total, subTotal) => total + subTotal, 0)
+    : 0;
   basket.account.consumptionTaxRate = 0.1;
   basket.account.totalItemsPrice = totalItemsPrice;
   const deliveryCharge = totalItemsPrice >= 5000 ? 0 : 500;
